reorganized tests + updated "contribute" docs + bumped celery and otxv2
