+++ requirements.txt
@@ -1,12 +1,12 @@
-amqp==2.5.1
+amqp==2.5.2
 asn1crypto==0.24.0
 atomicwrites==1.3.0
 attrs==19.1.0
 Babel==2.7.0
-billiard==3.6.1.0
+billiard==3.6.3.0
 boto3==1.9.216
 botocore==1.12.216
-celery==4.3.0
+celery==4.4.1
 certifi==2019.6.16
 cffi==1.12.3
 chardet==3.0.4
@@ -32,7 +32,7 @@ jmespath==0.9.4
 joblib==0.13.2
 jsbeautifier==1.6.2
 jsonschema==3.2.0
-kombu==4.6.5
+kombu==4.6.8
 lightgbm==2.1.2
 maxminddb==1.5.2
 more-itertools==7.2.0
@@ -40,7 +40,7 @@ msoffcrypto-tool==4.10.1
 numpy==1.17.1
 olefile==0.46
 oletools==0.55
-OTXv2==1.5.3
+OTXv2==1.5.8
 peepdf==0.4.2
 pefile==2019.4.18
 Pillow==3.2.0
